+++ dvr_scan/__init__.py
@@ -2,12 +2,12 @@
 #
 #      DVR-Scan: Video Motion Event Detection & Extraction Tool
 #   --------------------------------------------------------------
-#       [  Site: https://github.com/Breakthrough/DVR-Scan/   ]
-#       [  Documentation: http://dvr-scan.readthedocs.org/   ]
+#       [  Site: https://www.dvr-scan.com/                 ]
+#       [  Repo: https://github.com/Breakthrough/DVR-Scan  ]
 #
 # Copyright (C) 2014-2023 Brandon Castellano <http://www.bcastell.com>.
-# PySceneDetect is licensed under the BSD 2-Clause License; see the
-# included LICENSE file, or visit one of the above pages for details.
+# DVR-Scan is licensed under the BSD 2-Clause License; see the included
+# LICENSE file, or visit one of the above pages for details.
 #
 """``dvr_scan`` Module
 
@@ -42,7 +42,7 @@ import dvr_scan.opencv_loader as _
 
 # Top-level imports for easier access from the dvr_scan module.
 from dvr_scan.platform import init_logger
-from dvr_scan.scanner import ScanContext
+from dvr_scan.scanner import MotionScanner
 
 # Used for module/distribution identification.
 __version__ = '1.6-dev0'