+++ dvr_scan/detector.py
@@ -2,148 +2,113 @@
 #
 #      DVR-Scan: Video Motion Event Detection & Extraction Tool
 #   --------------------------------------------------------------
-#       [  Site: https://github.com/Breakthrough/DVR-Scan/   ]
-#       [  Documentation: http://dvr-scan.readthedocs.org/   ]
+#       [  Site: https://www.dvr-scan.com/                 ]
+#       [  Repo: https://github.com/Breakthrough/DVR-Scan  ]
 #
 # Copyright (C) 2014-2023 Brandon Castellano <http://www.bcastell.com>.
-# PySceneDetect is licensed under the BSD 2-Clause License; see the
-# included LICENSE file, or visit one of the above pages for details.
+# DVR-Scan is licensed under the BSD 2-Clause License; see the included
+# LICENSE file, or visit one of the above pages for details.
 #
-"""``dvr_scan.detector`` Module
+""" ``dvr_scan.detector`` Module
 
-Implementations of motion detectors. Currently, detectors use background subtraction to estimate the
-amount of motion in a sequence of frames. The frame score is calculated from how much of the frame
-contains foreground (moving) elements."""
+Contains the motion detection algorithm (`MotionDetector`) for DVR-Scan.  It calculates a score that
+represents the relative amount of movement of consecutive frames in a video.
+"""
 
-from abc import ABC, abstractmethod
+from collections import namedtuple
+from dataclasses import dataclass
+import logging
+import typing as ty
 
-import cv2
-import numpy
+import numpy as np
+
+from dvr_scan.subtractor import Subtractor
+
+Rectangle = namedtuple("Rectangle", ['x', 'y', 'w', 'h'])
+
+logger = logging.getLogger('dvr_scan')
 
 
-class Detector(ABC):
-    """Provides a consistent interface for calculating a mask of areas within each video
-    frame containing any motion."""
+@dataclass
+class ProcessedFrame:
+    """Result of processing a frame through the `MotionDetector`."""
 
-    @abstractmethod
-    def apply(self, frame: numpy.ndarray) -> numpy.ndarray:
-        """Apply the background subtractor to the given frame.
+    subtracted: np.ndarray
+    """Mask representing areas of cropped input frame that have motion."""
+    masked: ty.Union[np.ndarray, np.ma.MaskedArray]
+    """The background mask with the specified ROIs applied."""
+    score: float
+    """Score representing relative amount of motion in this frame inside the specified ROIs."""
 
-        Arguments:
-            frame: Frame to perform background subtraction on.
 
-        Returns:
-            Mask of areas in the frame containing motion.
-        """
+class MotionDetector:
+    """Detects motion on the input provided by the associated MotionScanner."""
+
+    def __init__(self, subtractor: Subtractor, frame_size: ty.Tuple[int, int], downscale: int,
+                 regions: ty.Optional[ty.Iterable[Rectangle]]):
+        logging.debug('frame size = %s', str(frame_size))
+        self._subtractor = subtractor
+        self._frame_size = frame_size
+        self._downscale = downscale
+        self._regions = list(regions) if not regions is None else []
+        self._mask: np.ndarray = np.ones((0, 0))
+        self._area: Rectangle = Rectangle(x=0, y=0, w=self._frame_size[0], h=self._frame_size[1])
+        logging.debug('%s', str(self._area))
+        if len(self._regions) > 1:
+            # For multiple ROIs, we calculate a mask array where True denotes ignored (masked)
+            # elements, and False represents unmasked elements - those inside the defined ROI.
+            min_x, min_y, max_x, max_y = self._frame_size[0], self._frame_size[1], 0, 0
+            for region in self._regions:
+                min_x, min_y = min(min_x, region.x), min(min_y, region.y)
+                max_x, max_y = max(max_x, region.x + region.w), max(max_y, region.y + region.h)
+            width, height = max(0, max_x - min_x), max(0, max_y - min_y)
+            logger.debug(f"creating mask, covers [{min_x},{min_y}] to [{max_x}, {max_y}]")
+            self._mask = np.ones((height, width), dtype=bool)
+            for region in self._regions:
+                region = Rectangle(x=region.x - min_x, y=region.y, w=region.w, h=region.h)
+                self._mask[region.y:region.y + region.h, region.x:region.x + region.w] = False
+            if self._downscale > 1:
+                self._mask = self._mask[::self._downscale, ::self._downscale]
+            self._area = Rectangle(x=min_x, y=min_y, w=width, h=height)
+        elif len(self._regions) == 1:
+            self._area = self._regions[0]
+
+    @property
+    def area(self) -> Rectangle:
+        """Area the region of interest covers in the original frame."""
+        return self._area
+
+    @property
+    def background_mask(self) -> np.ndarray:
         raise NotImplementedError()
 
-    @staticmethod
-    @abstractmethod
-    def is_available():
-        """True if this detector is available (e.g. has all dependencies), False otherwise."""
-        raise NotImplementedError()
-
-
-class DetectorMOG2(Detector):
-    """MOG2 background subtractor."""
-
-    def __init__(
-        self,
-        kernel_size: int,
-        history: int = 500,
-        variance_threshold: int = 16,
-        detect_shadows: bool = False,
-    ):
-        if kernel_size < 0 or (kernel_size > 1 and kernel_size % 2 == 0):
-            raise ValueError("kernel_size must be >= 0")
-        self._kernel = numpy.ones(
-            (kernel_size, kernel_size), numpy.uint8) if kernel_size > 1 else None
-        self._subtractor = cv2.createBackgroundSubtractorMOG2(
-            history=history,
-            varThreshold=variance_threshold,
-            detectShadows=detect_shadows,
-        )
-        # Default shadow value is 127, set to 0 so they are discarded before filtering.
-        self._subtractor.setShadowValue(0)
-
-    def apply(self, frame: numpy.ndarray) -> numpy.ndarray:
-        frame_gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-        frame_mask = self._subtractor.apply(frame_gray)
-        if not self._kernel is None:
-            frame_filt = cv2.morphologyEx(frame_mask, cv2.MORPH_OPEN, self._kernel)
+    def _preprocess(self, frame: np.ndarray) -> np.ndarray:
+        cropped = None
+        if not self._regions:
+            cropped = frame
+        elif len(self._regions) == 1:
+            cropped = frame[self._regions[0].y:self._regions[0].y + self._regions[0].h,
+                            self._regions[0].x:self._regions[0].x + self._regions[0].w]
         else:
-            frame_filt = frame_mask
-        return frame_filt
+            cropped = frame[
+                self._area.y:self._area.y + self._area.h,
+                self._area.x:self._area.x + self._area.w,
+            ]
+        if self._downscale > 1:
+            return cropped[::self._downscale, ::self._downscale, :]
 
-    @staticmethod
-    def is_available():
-        return hasattr(cv2, 'createBackgroundSubtractorMOG2')
+        return cropped
 
+    def update(self, frame: np.ndarray) -> ProcessedFrame:
+        frame = self._preprocess(frame)
+        subtracted = self._subtractor.apply(frame)
 
-class DetectorCNT(DetectorMOG2):
-    """CNT background subtractor."""
+        if len(self._regions) <= 1:
+            return ProcessedFrame(
+                subtracted=subtracted, masked=subtracted, score=np.average(subtracted))
 
-    def __init__(
-        self,
-        kernel_size: int,
-        min_pixel_stability: int = 15,
-        use_history: bool = True,
-        max_pixel_stability: int = 15 * 60,
-        is_parallel: bool = True,
-    ):
-        if kernel_size < 0 or (kernel_size > 1 and kernel_size % 2 == 0):
-            raise ValueError("kernel_size must be odd integer >= 1 or zero (0)")
-        self._kernel = numpy.ones(
-            (kernel_size, kernel_size), numpy.uint8) if kernel_size > 1 else None
-        self._subtractor = cv2.bgsegm.createBackgroundSubtractorCNT(
-            minPixelStability=min_pixel_stability,
-            useHistory=use_history,
-            maxPixelStability=max_pixel_stability,
-            isParallel=is_parallel,
-        )
-
-    @staticmethod
-    def is_available():
-        return hasattr(cv2, 'bgsegm') and hasattr(cv2.bgsegm, 'createBackgroundSubtractorCNT')
-
-
-class DetectorCudaMOG2(DetectorMOG2):
-    """CUDA-accelerated MOG2 background subtractor."""
-
-    def __init__(
-        self,
-        kernel_size: int,
-        history: int = 500,
-        variance_threshold: int = 16,
-        detect_shadows: bool = False,
-    ):
-        if kernel_size < 0 or (kernel_size > 1 and kernel_size % 2 == 0):
-            raise ValueError("kernel_size must be odd integer >= 1 or zero (0)")
-        self._filter = cv2.cuda.createMorphologyFilter(
-            cv2.MORPH_OPEN, cv2.CV_8UC1, numpy.ones(
-                (kernel_size, kernel_size), numpy.uint8)) if kernel_size > 1 else None
-        self._subtractor = cv2.cuda.createBackgroundSubtractorMOG2(
-            history=history,
-            varThreshold=variance_threshold,
-            detectShadows=detect_shadows,
-        )
-        # Default shadow value is 127, set to 0 so they are discarded before filtering.
-        self._subtractor.setShadowValue(0)
-
-    def apply(self, frame: numpy.ndarray) -> numpy.ndarray:
-        stream = cv2.cuda_Stream()
-        frame_rgb_dev = cv2.cuda_GpuMat()
-        frame_rgb_dev.upload(frame, stream=stream)
-        frame_gray_dev = cv2.cuda.cvtColor(frame_rgb_dev, cv2.COLOR_BGR2GRAY, stream=stream)
-        frame_mask_dev = self._subtractor.apply(frame_gray_dev, -1, stream=stream)
-        if not self._filter is None:
-            frame_filt_dev = self._filter.apply(frame_mask_dev, stream=stream)
-        else:
-            frame_filt_dev = frame_mask_dev
-        frame_filt = frame_filt_dev.download(stream=stream)
-        stream.waitForCompletion()
-        return frame_filt
-
-    @staticmethod
-    def is_available():
-        return hasattr(cv2, 'cuda') and hasattr(cv2.cuda, 'createBackgroundSubtractorMOG2')
+        motion_mask = np.ma.array(subtracted, mask=self._mask)
+        return ProcessedFrame(
+            subtracted=subtracted,
+            masked=motion_mask,
+            score=np.ma.sum(motion_mask) / float(np.ma.count(motion_mask)))